refactor: add use notification center hook
