+++ packages/notification-center/src/components/notification-center/components/layout/header/HeaderContainer.tsx
@@ -1,10 +1,10 @@
 import React, { useContext, useEffect } from 'react';
 import { Header } from './Header';
-import { NotificationCenterContext } from '../../../../../store/notification-center.context';
-import { UnseenCountContext } from '../../../../../store/unseen-count.context';
+import { UnseenCountContext } from '../../../../../store';
+import { useNotificationCenter } from '../../../../../hooks';
 
 export function HeaderContainer() {
-  const { onUnseenCountChanged, header } = useContext(NotificationCenterContext);
+  const { onUnseenCountChanged, header } = useNotificationCenter();
   const { unseenCount } = useContext(UnseenCountContext);
 
   useEffect(() => {
