+++ packages/notification-center/src/hooks/index.ts
@@ -8,3 +8,4 @@ export * from './use-api.hook';
 export * from './use-subscriber-preference.hook';
 export * from './use-novu-theme-provider.hook';
 export * from './use-screens.hook';
+export * from './use-notification-center-hook';
